Corrected 'len' variable from 'int' to 'ssize_t'.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13568 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXregistry/src/test/SipRedirectServerTest.cpp
@@ -100,7 +100,7 @@ public:
       
       // check sipfrag body
       const HttpBody* pBody;
-      int len;
+      ssize_t len;
       CPPUNIT_ASSERT( ( pTxt = response.getHeaderValue( 0, SIP_CONTENT_TYPE_FIELD ) ) != 0 );
       ASSERT_STR_EQUAL( "message/sipfrag", pTxt );
       CPPUNIT_ASSERT( ( pBody = response.getBody() ) != 0 );
@@ -211,7 +211,7 @@ public:
       
       // check sipfrag body
       const HttpBody* pBody;
-      int len;
+      ssize_t len;
       CPPUNIT_ASSERT( ( pTxt = response.getHeaderValue( 0, SIP_CONTENT_TYPE_FIELD ) ) != 0 );
       ASSERT_STR_EQUAL( "message/sipfrag", pTxt );
       CPPUNIT_ASSERT( ( pBody = response.getBody() ) != 0 );
